[Feat] 네번째 cell 데이터 띄우기 - MapKit

+++ OpenWeather/Model/CommonModel.swift
@@ -16,7 +16,7 @@ struct Clouds: Decodable {
 // MARK: - Coord
 struct Coord: Decodable {
     let lat: Double
-    let lon: Int
+    let lon: Double
 }
 
 // MARK: - Sys
